Update information in `DESCRIPTION` file
This replaces the default content of the `DESCRIPTION` file with something meaningful. Also GPLv3 is now set as the license of the package.
And the package was cleaned up a bit, which includes putting `.Rbuildignore` under version control, deleting `Read-and-delete-me`, as well as adding the imports to the `NAMESPACE` file.